Moving to react-redux
This is necessary because of the annoying amount of props I have to pass
down from parents to children. The selectedClass props must be seen by
almost every logical component and it is untenable from a design
perspective to continue to create messy code to accomodate that.

Hopefully redux will fix this with its state control functionality.