Fixing ResultPanel such that it does not rerender every time new props are added - that was a pretty glaring issue.
Fixing ResultPanel such that when a schedule fails to generate, it does
not matter how actions are called - a message will be displayed only
once.

Fixing ResultPanel such that generate does not need to be hit twice in
order to see the new schedule - this is due to reusing the same keys, so
just used an autoincrement field instead.

Added some minor tests, however more enzyme testing required for the
components.